[#105] Refactor: modify NewPlan to consist of small components
- 코드의 가독성을 위해 planName, workoutTime, workoutTable을 지정하는 각 코드들을 컴포넌트로 정의하여 합성해서 사용함
- semantic을 지키기 위해 컴포넌트의 가장 바깥 태그를 section 태그로 변경함
